userController login, serviceDetail 수정, 헤더 html, css, js 수정, serviceDetail 페이지 모달창 추가
