+++ washgogo/src/main/resources/templates/guide/guide.html
@@ -922,7 +922,6 @@
             <button type="button" class="modal-close"></button>
         </div>
     </div>
-    -->
     <!---- App-Footer ---->
     <nav th:replace="~{/fragments/headerFooter:: footer}">
     </nav>
