+++ washgogo/src/main/resources/templates/index.html
@@ -67,7 +67,7 @@
                     <a href="/priceSheet"><button type="button" class="btn btn-border">가격표 보기</button></a>
                 </div>
                 <div class="right-side">
-                    <a href="/serviceDetail">
+                    <a href="/user/serviceDetail">
                         <div class="choice1">
                             <div class="text">
                                 <h2>여러 번 이용</h2>
@@ -76,7 +76,7 @@
                             <img class="btn-main-service" th:src="@{/img/iconTwice.png}" alt="여러 번 이용">
                         </div>
                     </a>
-                    <a href="/serviceDetail">
+                    <a href="/user/serviceDetail">
                         <div class="choice2">
                             <div class="text">
                                 <h2>한 번 이용</h2>
