+++ washgogo/src/main/resources/templates/service/serviceDetail.html
@@ -111,10 +111,23 @@
                     <span>별도 케어에 따른 추가 비용은 발생하지 않으며,안심 정찰 가격표에 의해 세탁 요금이 부과됩니다.</span>
                     <a href="/priceSheet">가격표 전체 보기 ></a>
                     <div class="button-area">
-                        <button type="button" onclick="subscribe()">서비스 신청하기</button>
+                        <button type="button" th:if="${userServiceType == null}" onclick="subscribe()">서비스 신청하기</button>
+                        <button type="button" th:if="${userServiceType == '여러 번 이용'}" onclick="modal()">서비스 신청하기</button>
                     </div>
                 </div>
             </div>
+        </div>
+    </div>
+    <div class="modal-bg"></div>
+    <div class="modal-container dialKeyInfo key">
+        <div class="modal-inner">
+            <h2 class="modal-title">이미 이용중인 서비스 입니다.</h2>
+            <div class="modal-content">
+                <div class="btn-area">
+                    <button type="button" class="btn-black">확인</button>
+                </div>
+            </div>
+            <button type="button" class="modal-close"></button>
         </div>
     </div>
     <!---- App-Footer ---->
@@ -134,8 +147,37 @@
         }
     }
 
-    function subscribe(){
-        location.href="/user/serviceSubscribeAddress";
+    const $modalBg = $(".modal-bg");
+    const $body = $("body");
+    const $closeModal = $(".modal-close");
+    const $okModal = $(".modal-container .btn-black");
+    const $modal = $(".modal-container");
+
+    /* 서비스 신청 버튼 이벤트 */
+
+    function subscribe() {
+        location.href = "/user/serviceSubscribeAddress";
+    }
+
+    function modal(){
+        $modal.css("display", "block");
+        modalOn();
+    }
+    /* 모달창 이벤트 */
+    $closeModal.on("click", function(){
+        modalOff();
+    });
+    $okModal.on("click", function(){
+        modalOff();
+    });
+    function modalOn(){
+        $modalBg.css("display", "block");
+        $body.css("overflow", "hidden");
+    }
+    function modalOff(){
+        $modalBg.css("display", "none");
+        $modal.css("display", "none");
+        $body.css("overflow", "auto");
     }
 </script>
 </html>
